progress on model-selection
- extended the RecommenderRequestBody class
- implemented a prototype of RecommenderService

+++ core/src/main/scala/at/magiun/core/model/data/DatasetMetadata.scala
@@ -6,6 +6,6 @@ case class DatasetMetadata(variableTypes: Seq[VariableType],
                            variableDistributions: Seq[Distribution],
                            responseVariableIndex: Int,
                            variablesToIgnoreIndex: Seq[Int],
-                           variablesCount: Int,
-                           observationsCount: Int) {
+                           variablesCount: Long,
+                           observationsCount: Long) {
 }
